+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v2` branch to see the changelog of Lume 2.
 Go to the `v1` branch to see the changelog of Lume 1.
 
-## [3.0.5] - Unreleased
+## [3.0.5] - 2025-07-17
 ### Changed
 - Unify browsers support accross several plugins like `postcss`, `highlightningcss` or `esbuild`.
 - Remove access request to env variables [#765]
@@ -175,7 +175,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 [#761]: https://github.com/lumeland/lume/issues/761
 [#765]: https://github.com/lumeland/lume/issues/765
 
-[3.0.5]: https://github.com/lumeland/lume/compare/v3.0.4...HEAD
+[3.0.5]: https://github.com/lumeland/lume/compare/v3.0.4...v3.0.5
 [3.0.4]: https://github.com/lumeland/lume/compare/v3.0.3...v3.0.4
 [3.0.3]: https://github.com/lumeland/lume/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/lumeland/lume/compare/v3.0.1...v3.0.2